DocumentImpl template params are used as encapsulated properties
Previously DocumentImpl was taking 4 template params, the two last ones for implementation (internal use).
This was used to have a writable property internally and expose only readonly types.
We are now going to expose writable copies so there is no need to separate internal and public types.
Writable copies will be more useful to users as they will be able to change the returned object directly.
They will have to pass this copied object back to the document so it can save the changes.